refactor CLI admin SDK functions
